other several changes to support multiple boards. each entity must have decoupled id to easen navigation. Huge impact

+++ src/main/java/it/angelic/growlroom/model/repositories/BoardsRepository.java
@@ -6,6 +6,6 @@ import it.angelic.growlroom.model.Board;
 
 public interface BoardsRepository extends JpaRepository<Board, Long> {
 
-	Board findByBoardId(int id);
+	Board findByBoardId(Long id);
  
 }
